LayoutTests/ipc/stream-sync-crash-no-timeout-expected.txt

CONSOLE MESSAGE: TypeError: Bad type name

FAIL Test that stream sync message crash is noticed immediately TypeError: Bad type name
